Cleaned the extracted code
