+++ src/components/extractor-result-model.tsx
@@ -1,7 +1,6 @@
 export interface ExtractionResult {
     success: boolean;
     rawText: string;
-    formattedSections?: { title: string; content: string }[];
     url?: string;
     title?: string;
     extractedFrom?: string | null;
