--- try2/background.js
@@ -1,5 +1,0 @@
-chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
-    console.log('Background received:', message);
-    sendResponse({status: 'Message received'});
-  });
-
