--- try2/manifest.json
@@ -1,21 +1,0 @@
-{
-    "manifest_version": 3,
-    "name": "Content Extraction Extension",
-    "version": "1.0",
-    "description": "Extract content from any URL using custom extraction code.",
-    "permissions": [
-      "scripting",
-      "activeTab"
-    ],
-    "host_permissions": [
-      "*://*/*"
-    ],
-    "action": {
-      "default_popup": "popup.html",
-      "default_title": "Extract Content"
-    },
-    "background": {
-      "service_worker": "background.js"
-    }
-  }
-
