--- try2/popup.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Content Extraction</title>
-    <style>
-      body {
-        font-family: Arial, sans-serif;
-        padding: 10px;
-        width: 300px;
-      }
-      #output {
-        white-space: pre-wrap;
-        border: 1px solid #ccc;
-        padding: 10px;
-        margin-top: 10px;
-        max-height: 300px;
-        overflow-y: auto;
-      }
-      input, button {
-        width: 100%;
-        margin: 5px 0;
-      }
-    </style>
-    <!-- Optionally, include PDF.js if you plan to extract PDFs -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/pdf.js/2.10.377/pdf.min.js"></script>
-  </head>
-  <body>
-    <h1>Extract Content</h1>
-    <input type="text" id="urlInput" placeholder="Enter URL" />
-    <button id="extractBtn">Extract</button>
-    <div id="output">Output will appear here.</div>
-    <script src="popup.js"></script>
-  </body>
-</html>
